+++ src/main/java/com/aetherteam/aether/item/tools/holystone/HolystoneAxeItem.java
@@ -9,7 +9,7 @@ import net.minecraft.world.item.Item;
  * Ambrosium dropping behavior is called by {@link com.aetherteam.aether.item.tools.abilities.ToolAbilities#handleHolystoneToolAbility}.
  */
 public class HolystoneAxeItem extends AxeItem implements HolystoneTool {
-    public HolystoneAxeItem() {
-        super(AetherItemTiers.HOLYSTONE, 7.0F, -3.2F, new Item.Properties());
+    public HolystoneAxeItem(Item.Properties properties) {
+        super(AetherItemTiers.HOLYSTONE, 7.0F, -3.2F, properties);
     }
 }